+++ NuKeeper.Inpection.Tests/NuGetApi/BulkPackageLookupTests.cs
@@ -8,7 +8,7 @@ using NuGet.Versioning;
 using NuKeeper.Inspection.NuGetApi;
 using NUnit.Framework;
 
-namespace NuKeeper.Tests.NuGet.Api
+namespace NuKeeper.Inpection.Tests.NuGetApi
 {
     [TestFixture]
     public class BulkPackageLookupTests
