+++ NuKeeper.Inpection.Tests/NuGetApi/PackageLookupResultReporterTests.cs
@@ -6,7 +6,7 @@ using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.NuGetApi;
 using NUnit.Framework;
 
-namespace NuKeeper.Tests.NuGet.Api
+namespace NuKeeper.Inpection.Tests.NuGetApi
 {
     [TestFixture]
     public class PackageLookupResultReporterTests
