+++ NuKeeper.Inpection.Tests/RepositoryInspection/PackageAssert.cs
@@ -1,7 +1,7 @@
 using NuKeeper.Inspection.RepositoryInspection;
 using NUnit.Framework;
 
-namespace NuKeeper.Tests.RepositoryInspection
+namespace NuKeeper.Inpection.Tests.RepositoryInspection
 {
     public static class PackageAssert
     {
