+++ NuKeeper.Inpection.Tests/RepositoryInspection/PackagesFileReaderTests.cs
@@ -5,7 +5,7 @@ using NuGet.Versioning;
 using NuKeeper.Inspection.RepositoryInspection;
 using NUnit.Framework;
 
-namespace NuKeeper.Tests.RepositoryInspection
+namespace NuKeeper.Inpection.Tests.RepositoryInspection
 {
     [TestFixture]
     public class PackagesFileReaderTests
